backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
sqlalchemy
pydantic
paho-mqtt
aiocoap
pytest
httpx

